+++ include/draw.h
@@ -9,7 +9,7 @@
 
 void draw_pixel(uint32_t *buffer, uint32_t x, uint32_t y, uint32_t color);
 void draw_raycast(uint32_t *buffer, Map* map,
-    double posX, double posY, double dirX, double dirY,
-    double planeX, double planeY);
+    float posX, float posY, float dirX, float dirY,
+    float planeX, float planeY);
 
 #endif
